Links機能の実装完了

+++ internal/api/model/user.go
@@ -1,13 +1,13 @@
 package model
 
 type User struct {
-	ID  	int64  `json:"id"`
+	ID  	uint  `json:"id"`
 	Name 	string `json:"name"`
 	Email 	string `json:"email"`
 	Password string `json:"password"`
 }
 
 type UserResponse struct {
-	ID  	int64  `json:"id"`
+	ID  	uint  `json:"id"`
 	Email 	string `json:"email"`
 }
